Mount volume to inspect config.json
Change-type: patch

+++ balena.sh
@@ -1,6 +1,21 @@
 #!/usr/bin/env bash
 
-set -e
+set -ex
+
+device="$(blkid | grep resin-boot | awk -F':' '{print $1}')"
+
+function cleanup() {
+   (sync && umount "${1}") || true
+}
+
+trap 'cleanup ${device}' SIGINT SIGTERM USR1 USR2 EXIT
+
+function mount_boot() {
+    local tmpmnt
+    tmpmnt="$(mktemp -d)"
+    mount "${1}" "${tmpmnt}"
+    echo "${tmpmnt}"
+}
 
 metadata_urls=(
   'http://169.254.169.254/latest/user-data'
@@ -12,16 +27,6 @@ curl_with_opts() {
     curl --fail --silent --connect-timeout 3 "$@"
 }
 
-ssh_with_opts() {
-    ssh -p 22222 \
-      -o 'StrictHostKeyChecking=no' \
-      -o 'UserKnownHostsFile=/dev/null' \
-      -o 'PasswordAuthentication=no' \
-      -o 'ConnectTimeout=60' \
-      "root@$(ip route | awk '/balena0|br-[0-9a-fA-F]/ { print $7 }' | head -n 1)" \
-      "$@"
-}
-
 config_from_metadata() {
     #shellcheck disable=SC2034,SC2039 # /bin/sh is a symbolic link to bash on balenaOS
     for url in "${metadata_urls[@]}"; do
@@ -30,34 +35,24 @@ config_from_metadata() {
     done
 }
 
-if ssh_with_opts -t id; then
-    # (legacy) requires SSH private key to be preloaded in the image and corresponding
-    # public key injected info config.json on the host OS
-    if [[ "$(ssh_with_opts "cat /mnt/boot/config.json | jq -r .deviceApiKey")" =~ null|^$ ]]; then
-        ssh_with_opts "os-config join '$(config_from_metadata)'"
-    fi
-else
-    # open fleet flow requires a reboot
-    tmpmnt="$(mktemp -d)"
-    device="$(blkid | grep resin-boot | awk -F':' '{print $1}')"
-    mount "${device}" "${tmpmnt}"
+tmpmnt="$(mount_boot "${device}")"
+tmpconf="$(mktemp)"
+config_from_metadata > "${tmpconf}"
 
-    tmpconf="$(mktemp)"
-    config_from_metadata > "${tmpconf}"
-
-    if [[ -f "${tmpconf}" ]]; then
-        app_id="$(cat < "${tmpconf}" | jq -r .applicationId)"
-
-        if [[ -n "${app_id}" ]]; then
-            if [[ "${RESIN_APP_ID}" != "${app_id}" ]]; then
-                cat < "${tmpconf}" > "${tmpmnt}/config.json" \
-                  && sync \
-                  && umount "${device}" \
-                  && curl -sX POST "${BALENA_SUPERVISOR_ADDRESS}/v1/reboot?apikey=${BALENA_SUPERVISOR_API_KEY}" \
-                  --header 'Content-Type:application/json'
-            fi
-        fi
-    fi
+if [[ -f "${tmpmnt}/config.json" ]] && [[ -f "${tmpconf}" ]]; then
+    device_api_key="$(cat < "${tmpmnt}/config.json" | jq -r .deviceApiKey)"
+    if [[ "${device_api_key}" =~ null|^$ ]]; then
+		cat < "${tmpconf}" > "${tmpmnt}/config.json"
+	else
+		app_id="$(cat < "${tmpconf}" | jq -r .applicationId)"
+		if [[ -n "${app_id}" ]]; then
+			if [[ "${RESIN_APP_ID}" != "${app_id}" ]]; then
+				cat < "${tmpconf}" > "${tmpmnt}/config.json" \
+				  && curl -sX POST "${BALENA_SUPERVISOR_ADDRESS}/v1/reboot?apikey=${BALENA_SUPERVISOR_API_KEY}" \
+				  --header 'Content-Type:application/json'
+			fi
+		fi
+	fi
 fi
 
 exec balena-idle "$@"
